pinky: move unit test away from integration tests
